+++ station/src/oskar_evaluate_tapered_dipole_pattern.cu
@@ -36,7 +36,7 @@ extern "C"
 #endif
 int oskar_evaluate_tapered_dipole_pattern(oskar_Mem* pattern,
         const oskar_Mem* theta, const oskar_Mem* phi, int cos_power,
-        int gaussian_fwhm_rad, int return_x_dipole)
+        double gaussian_fwhm_rad, int return_x_dipole)
 {
     int type, num_sources;
